+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 0.1.0 (In Development)
 ----------------------
 
+0.1.0a3 (In Development)
+========================
+
+- Allow passing dict of URLs with list methods to flex process of matching
+  request ignored to wrapping into timeout context manager
+
 0.1.0a2 (2017-05-14)
 --------------------